fix: the issue where the Web component fails to initialize when an error occurs during hot reload.
Co-authored-by: 谢代荣<xiedairong@huawei.com>



# message auto-generated for no-merge-commit merge:
merge master-webview into master

fix: the issue where the Web component fails to initialize when an error occurs during hot reload.

Created-by: xiedrsz
Commit-by: 谢代荣
Merged-by: huangyouhua
Description: ## Description
When using Metro to load the jsBundle, if an error occurs on the JavaScript side, it will trigger the fetchCookies method in httpClient. This method initializes the web engine. If the app has enabled worker threads, the web engine will be initialized in the worker thread, which can cause the initialization of the web engine in the main thread to fail. This PR fixes the issue by initializing the web engine in advance within the httpClient constructor.

## Changes
- fixed the issue where the Web component fails to initialize when an error occurs during hot reload.

## Test Plan
1. Open the Tester Project.
1. Replace `ContentSlot(this.ctx.getContentForTag(this.tag))` in ContainerView.ets with `Web({ src: 'www.baidu.com', controller: this.controller })`.
1. Check whether the ContainerView component in the CustomNativeComponent test case displays the webpage properly.

## Checklist
- [x] Does not involve incompatible changes; if involved, has been reviewed accordingly.
- [x] Does not impact performance, or performance testing has been conducted without degradation.
- [x] Complies with the relevant coding standards.
- [x] Does not involve documentation updates, or the documentation has been updated.
- [x] Meets testability requirements with necessary self-test cases, appropriate logging, or trace information added.
- [x] No illegal file inclusions exist, such as images or code.


See merge request: openharmony-sig/ohos_react_native!1504