fix: Cursor abnormally resets inside TextInput component
Co-authored-by: jiaxuan_wong<1801330925@qq.com>



# message auto-generated for no-merge-commit merge:
merge master into master

fix: Cursor abnormally resets inside TextInput component

Created-by: jiaxuan_wong
Commit-by: jiaxuan_wong
Merged-by: huangyouhua
Description: 
**Description**
When typing inside a TextInput that is placed in a Modal, the cursor occasionally jumps to the beginning after the Modal is closed and reopened.
Root cause:
When the Modal is reopened, m_selectionLocation is re-initialized to 0, so all subsequent “distance-to-end” calculations start from the wrong origin.
The helper function countUtf16Characters() under-counts the length of strings that contain emoji (because it does not take surrogate-pairs into account), so the cursor is positioned in the middle of the text instead of at the end.
**Change Log**

-countUtf16Characters()
-m_selectionLocation

**Test Steps**

1.Open the Tester project.
2.Run npm run start.

**Pre-Merge Checklist**
<!--  
Please perform a self-check according to the items below.  
Put an “x” in [ ] after you finish each item, e.g.  
- [x] No incompatible changes, or already reviewed if any.  
-->  
[ ] No incompatible changes; if any, relevant review passed.
[x] No performance impact, or performance tested without regression.
[x] Follows coding standard.
[ ] No documentation update required, or documentation already updated.
[ ] Necessary test cases, logging or trace information added for testability.
[x] No illegal files (images, code, etc.) introduced.
[ ] This PR has been opened against branch 77, or does not need to be merged into branch 77.
 

See merge request: openharmony-sig/ohos_react_native!1568