fix: Add the launchParam parameter in the onCreate of RNAbility.
Co-authored-by: saberblack<2422489333@qq.com>



# message auto-generated for no-merge-commit merge:
!1893 merge master into master

fix: Add the launchParam parameter in the onCreate of RNAbility.

Created-by: saberblack
Commit-by: saberblack
Merged-by: huangyouhua
Description: <!-- PLEASE REMOVE COMMENTS BEFORE MERGING TO AVOID INCLUDING THEM IN GIT HISTORY -->
## Description
<!--
Briefly describe the purpose of the MR, for example:
This PR mainly implements the function of remote debugging within the Metro LAN.
-->
Add the launchParam parameter in the onCreate of RNAbility.

## Changes
<!--
List main changes in bullet points that can be included in the changelog. Those changes should be written for
application developers and react third-party library authors. Internal changes shouldn't be listed. Please start each bullet point with:
- added
- fixed
- deprecated
- removed

If it is a change to add a new attribute, use [added] prefix, for example:
- [added] FlatList component adds inverted attribute
-->
1. Add the launchParam parameter in the onCreate of RNAbility.
2. Edit the related README in English.

## Test Plan
<!-- 
How to test this PR. For example:
1. Open the Tester project
2. Execute the npm run start command
...
-->



## Checklist
<!--
Please conduct self inspection according to the following checklist before merging
After completing the self check, fill in "x" in [], For example:
- [x] Does not involve incompatible changes; if involved, has been reviewed accordingly.
-->
- [x] Does not involve incompatible changes; if involved, has been reviewed accordingly.
- [x] Does not impact performance, or performance testing has been conducted without degradation.
- [x] Complies with the relevant coding standards.
- [x] Does not involve documentation updates, or the documentation has been updated.
- [x] Meets testability requirements with necessary self-test cases, appropriate logging, or trace information added.
- [x] No illegal file inclusions exist, such as images or code.


See merge request: openharmony-sig/ohos_react_native!1893